AI Image Detector/requirements.txt
customtkinter>=5.0.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
